Seed the bosh-init image with compiled AWS CPI.
Compiling the bosh AWS CPI takes c. 4 minutes every time bosh-init is
run.

This adds a dummy deployment to the image build process, which causes
the AWS CPI to be compiled then, and written to root's homedir. This is
then shipped as part of the image, and is therefore available for all
runs of the container.

This will only have an effect when deploying manifests that use a
matching version of the AWS CPI. It's therefore necessary to keep the
version (specified by the BOSH_AWS_CPI_URL and BOSH_AWS_CPI_CHECKSUM
variables in the Dockerfile) in sync.