Seed the installation_id for the bosh deploy.
This id is used in the cache directory structure. It's therefore
necessary to seed it both here, and for the real bosh-init runs, using
the same value in order for the cached compilations to be picked up.

The actual UUID used here is just an arbritrary value. The specific
value doesn't matter, it just has to be the same in all places.