Revert "Container name change to math our offcial one"
This reverts commit 72bb2e4d575c65a23540b03b0893ed481ec0722d.

@mtekel noted in #10 that the instructions in the README for building
currently conflict with being able to run the tests locally.

I think we should use a temporary name, which doesn't match the Docker Hub
repo name, when building the container locally and in Travis for testing.
This reduces the chance of confusing it with the "real" container that
you've pulled from the public registry.

It also matches the behaviour of the containers in:

- alphagov/paas-docker-cloudfoundry-tools