Test for size of binary directory
In order to document and enforce the intent of whitelisting binaries in the
previous commit. The current installation of 0.6.10 is 144M, so 200M should
allow for variations in later releases, until we add other providers.

I tried using serverspec's `File#size` matcher, but that only gives the size
of the directory node (4096 bytes) rather than its contents.