Whitelist binaries to reduce container size
The Terraform installation is quite large; 30~ provider and provisioner
binaries at 10~20M each. In order to reduce the size of the final container
we only install the providers and provisioners that we expect to use. A list
of these is maintained in the [Dockerfile](Dockerfile). We don't expect to
need to do this in the future if hashicorp/terraform#4233 is solved.

I've picked all of the non-provider specific binaries with the exception of
AWS which we currently deploy alphagov/paas-cf to.

I did look at using upx, as described in the upstream issue. However it's
not easy to make use of it in our current build process. We'd have to
include Go to build goupx and besides Alpine Linux doesn't have a upx
package to build against. So we'd have to create/use a larger container just
to use it.

This brings down the size of the container from 590M to 160M:

    ➜  ~  docker images | grep terraform
    terraform                         latest              7441268d0c1c        About an hour ago   162.1 MB
    governmentpaas/docker-terraform   latest              f8a6bada0c7a        3 hours ago         590.8 MB

As seen from within the container:

    ➜  ~  docker run -ti --rm terraform du -h /usr/local/bin
    144.9M  /usr/local/bin
    ➜  ~  docker run -ti --rm governmentpaas/docker-terraform du -h /usr/local/bin
    553.8M  /usr/local/bin