Add cf-cli image.
This contains the CF CLI, and a Go runtime. This will be used to run the
acceptance tests in concourse.

We looked at using the 1.5.3-alpine Go image as a base, but the
published cf-cli releases are compiled against glibc and don't work with
the musl C library that apline uses.