Update bosh-cli openssl version.
httpclient (which is a dependency of bosh-cli) has a check for the
openssl version[1] to ensure it includes fixes for a vulnerability.
Versions below this cause it to emit warnings.

Because debian backports security fixes instead of upgrading openssl,
these warnings are triggered here, even though openssl is no longer
vulnerable.

Digging further, the CVE mentioned in the comment in httpclient
specifies that the only versions affected are 1.0.1n, 1.0.1o, 1.0.2b and
1.0.2c, so the httpclient check identifies non-vulnerable versions as
vulnerable.

To work around this, we've switched the bosh-cli container to be based
on alpine, which includes the newer versions of openssl.

[1] https://github.com/nahi/httpclient/blob/v2.7.1/lib/httpclient/ssl_config.rb#L441-L452