Use the latest spruce 1.1.2 from release
We have been using the version 0.13.0 of spruce, but compiling it
in the Dockerfile because the release binary version did link to
the debian/ubuntu libc. We were not pinning the git commit of the
spruce code.

But the upstream project has some changes that makes fail the build,
probably new dependencies added which require the use of godep.

As since https://github.com/geofffranks/spruce/issues/83 the binary
in the releases is OK and can work in a plain alpine, we can create
a container pulling the binary release. This will simplify our
container, make it smaller and pin the version number.