TF: build from tag
Build TF inside container from the tag specified in the Dockerfile.
This way we can easily fast forward to yet unreleased master/tags,
which contain fixes/new features we need. Once these are released
as proper version, we can change the tag from commit to the version
number. TF uses v0.6.15 scheme.

The tag in this commit is for 0.6.15-dev version. Reflect this in
tests.