Instal awscli in self-update-pipelines container
For the setup of the pool resource for the pipeline locking, we need
to download some objects from S3. We decided use aws cli for this.

In this commit we install awscli from pip in ubuntu, using pip to
easily get the latest version.

We need to install also groff+less to have an operational help.