Revert "[#116983391] Bump cf-cli versions to meet reqs for CF v233"
In the new version of the client they changed the behaviour of the `-d`
option for `cf curl`, and it does not allow pass an empty string.

As our pipelines always pull the latest container version from the
respository, this PR makes the current acceptance tests
(pinned by current cf-release v228) fail, as they pass a empty string
to `-d` (see [1])

I revert the version upgrade to keep compatibility with the current
release used in github.com/alphagov/paas-cf:master.

For the update, I suggest that we update the client after upgrading
the release in all the environments, and we use, meanwhile,
tagged containers in the cf-acceptance tests in order to correlate
the acceptance tests with the client version.

More info:
[1] https://github.com/cloudfoundry-incubator/cf-test-helpers/commit/be97699700d8615a9b005065c34ae8ec981795bc
[2] https://github.com/cloudfoundry/cli/commit/b2ca2d045f79f260339798ae3a4878f1ed31d1e8