Install bash in the git-ssh container
We use fair complex scripts with these containers, on which we want
to set options like pipefail [1]. Alpine `sh` does support that option
but not the GNU sh, which causes the scripts fail if they are
executed or tested outside of this container (e.g. travis).

To solve this, we install bash in the container so that we can use
it in the scripts and make them portable.

[1] https://bclary.com/blog/2006/07/20/pipefail-testing-pipeline-exit-codes/