Terraform: Validate checksum of downloaded release
To ensure that every build gets the exactly the same version/release that we
have specified.

This will guard against the unlikely events that the release has been
updated without changing the version number (accidentally or maliciously) or
our download has been MITMed.

The checksum has been taken from:

- https://releases.hashicorp.com/terraform/0.6.13/terraform_0.6.13_SHA256SUMS

This isn't perfect, as described in the following article, but it's still
good practice:

- https://micahflee.com/2016/02/backdoored-linux-mint-and-the-perils-of-checksums/