Terraform: Upgrade to 0.6.15
This isn't as big a jump as this individual commit suggests because we were
already using a pre-release version of 0.6.15 in 57cc1a9. We are upgrading
so that we no longer need to build from source to benefit from these fixes:

- hashicorp/terraform#5774
- hashicorp/terraform#5904

The CHANGELOG doesn't mention any backwards-incompatible changes. New
checksum has been taken from:

- https://releases.hashicorp.com/terraform/0.6.15/terraform_0.6.15_SHA256SUMS