Add json-minify container
It is based on Ruby and adds the json-minify gem.

We require use of the json-minify Ruby Gem to insert JSON as values
in our CloudFoundry manifest. Each of these lines represents the
configuration of a Grafana dashboard. This image allows us to isolate
this dependency and not have to run bundler in the pipeline.