Fix bosh-cli wget to https sites.
The current version of the ruby:2.2-alpine no longer includes the
openssl CLI. That means that the busybox provided wget command can no
longer connect to https sites, and instead gives the following error:

wget: can't execute 'ssl_helper': No such file or directory
wget: error getting response: Connection reset by peer

This adds openssl to this container to fix this issue.

This partially reverts cfc7785e, which updated the tests to check for
the libssl version instead because we didn't think that the openssl CLI
was needed.