cf-cli: Add jq 1.5 with regex support
We need to parse the output of `cf curl /v2/users` to list all the
existing users created by the acceptance tests. For this jq is perfect.

We consider that this really useful and it will be used in other
stories in the future, so we add it as a basic tool in the cf-cli
container.

We want to use the regex based functions, so jq must be >= 1.5 and must
be compiled with ONIGURUMA regex library. The Ubuntu package is 1.4, so
we will pull the static binary released upstream.