cf-cli: Use alpine base system for cf-cli container
Alpine base image[1] is much more lightweight than a normal Ubuntu. We
use that base image whenever is possible to speed up the pipelines.

But in the case of cf-cli we were using Ubuntu because the cf-cli was
not compiled statically. Since 6.17.1. they build it statically[2], so
we can move the container to alpine.

[1] https://hub.docker.com/_/alpine/
[2] https://github.com/cloudfoundry/cli/issues/807