cf-cli: Install ruby gem json in cf-cli container
After merging #62, our pipeline fail when executing a ruby script
which required json.

Although json is one of the basic packages shipped with ruby[1],
Alpine extracts it in a different package ruby-json. We need to
explicitely add that dependency.

We avoid running `bundle install` in the pipeline to avoid
increase the time to run. We normally try to keep the scripts simple
and without dependencies, but `json` is a really basic library.

[1] https://zzz.buzz/2016/02/29/default-gems-bundled-with-ruby/
[2] http://git.alpinelinux.org/cgit/aports/tree/main/ruby/APKBUILD?h=3.4-stable&id=cbdb9b2e42bbd7a67f0945bf7bcb8e79ddcae742#n190