Create enum-resource for storing an enumerated value in S3
The user provides a set of allowed values. The resource mediates access to the storage of a single value, ensuring that only allowed values are available to the task.

The approach is to base it on [Smuggler](https://github.com/redfactorlabs/concourse-smuggler-resource/tree/master/examples/s3-with-default).
This allows us to define the `check`, `in` and `out` in terms of simple shell scripts.

We define a docker container to provide the resource to the pipeline manifest.

We use the (paas-s3-resource)[https://github.com/alphagov/paas-s3-resource] internally to access S3.

There's a `README` and an example pipeline showing the expected usage.

This could serve as an example for future smuggler-derived concourse resources.