cf-cli: Install libc6-compat
We want to use a blue-green-deploy plugin but it is currently distributed as
a dynamically linked binary which doesn't work in this container because
Alpine uses musl-libc.

The `libc6-compat` package provides `ld-linux-x86-64.so.2` which enables
glibc linked binaries to work. Another commit will follow to make this work
for 64bit binaries on Alpine <= 3.5

We may be able to revert this in the future if the plugin is released as a
statically linked binary as requested in:

- bluemixgaragelondon/cf-blue-green-deploy#28