cf-cli: Install blue-green-deploy plugin
We want to use this plugin within our alphagov/paas-cf pipeline. At the
moment it will only be for the healthcheck app, which we never want to have
any downtime because we use it for monitoring and automated tests. We may
choose to use it for other apps in the future if this proves successful.

We chose to use a plugin instead of scripting the map commands by ourselves
because we don't want to deal with any error handling or edge cases that may
not be immediately apparent at this time. We chose this plugin over
contraband/autopilot because it seems to be more actively maintained and
tested.

The `cf install-plugin` command doesn't currently allow you to specify the
version of the plugin. This is important to us because we don't want the
version to change unexpectedly when we rebuild the container, which can
happen quite frequently. I've raised a feature request to discuss this in
cloudfoundry/cli#1067

I've added checksum validation because this plugin will have access to CF
credentials and we would want to know if the version hosted on GitHub
releases has been modified since we first started using it.

The chmod is required otherwise this happens:
```
Installing plugin blue-green-deploy.linux64...
FAILED
fork/exec /tmp/blue-green-deploy.linux64: permission denied
```