Tag each Docker build with git commit ref
We want every successful build of a Docker image to have a unique tag that
we can reference in our Concourse pipelines so that we can control the
upgrades of new versions to different environments (e.g. promoting from
staging to production).

This uses Docker Cloud/Hub custom build hooks. The code is actually taken
from an example in their documentation:

- https://docs.docker.com/docker-cloud/builds/advanced/#push-to-multiple-tags

The documentation refers to Docker Cloud, which is the ongoing replacement
of Docker Hub that we currently use. I think both may share some of the same
underlying infrastructure. Nonetheless custom build hooks are supported by
Docker Hub too.

I've added an `echo` to help debug it if there are any problems in the
future. This appears in the "build details" log in Docker Hub.