Ensure that $SOURCE_COMMIT is set in build hook
It seems that this variable isn't set for Docker Hub builds. This might be
a discrepancy between Docker Hub and Docker Cloud. It doesn't matter though,
because we have access to `git` and can construct the value ourselves if it
isn't set. This will be a no-op if it gets fixed in the future.