New bosh-shell docker image
This is intended for use as an interactive shell by an operator on
their local machine. This differs from the bosh-cli, which is intended
for use in our pipelines.

Because this is an interactive container we provide an `ENTRYPOINT`
script which sets up a local port forward of the Bosh port (25555)
and logs into Bosh. Operators are still required to `bosh download
manifest` and target a deployment.

We pass secrets in at runtime to avoid them getting baked into the image.

We allow you to run the container without setting `BOSH_ID_RSA` and
`BOSH_IP` because you may need to interact with the container without
setting all that stuff up, and if we don't allow it the RSpec tests
won't work, as they rely on interacting with the container and they
can't access the secrets from Travis.