bosh-shell: Automatically target BOSH deployment
So that you don't have to run these commands yourself each time you use
`make bosh-cli` from paas-bootstrap, which I find very frustrating. The old
target in paas-cf used to do an equivalent of this for you.

This requires a new environment variable to be passed from paas-bootstrap.
I've called it something other than `DEPLOY_ENV` to make it easier to
understand what it is used for.