terraform: Bake in provider plugins
Recent versions of Terraform no longer ship with the providers built into
the main `terraform` binary. They are downloaded based on the `provider{}`
blocks that you have in your `*.tf` codebase.

We want to bake these into the container so that each run in the pipeline
doesn't need to download them from the internet because it will a) save
bandwidth and b) prevent the providers being upgraded automatically meaning
that two runs of the same job should have the same effect.

We experimented with using Terraform's plugin cache option, which stores the
download in the separate directory and uses hardlinks for every subsequent
`terraform init` when the environment variable is present. However if a new
version of a plugin is published then it will be automatically downloaded
from the Internet instead of using the cache:

- https://www.terraform.io/docs/configuration/providers.html#provider-plugin-cache

So the solution we're using instead is baking the `.terraform` directory
that we used to first download the plugins into the container and setting
the `TF_DATA_DIR` environment variable so that all subsequent `terraform`
runs also use it. This directory contains a lock file which with the default
of `-upgrade=false` will prevent the plugins from changing version. Re-using
the data directory could have bad effects if you were running multiple
Terraform codebases from the same container but that's not a problem for us
because each invocation will use a new instance of the container.

- https://www.terraform.io/docs/configuration/environment-variables.html#tf_data_dir