terraform: Disable some prompts for interactive input
This commit sets the `TF_INPUT` environment variable to `0`. This is
documented as "used when you want to disable prompts for variables
that haven't had their values specified" [1]. Setting this in the
container is less transparent than in the pipeline, but ensures fewer
demands for interactive input in our pipelines.

There is no single means for making Terraform run non-interactively.
We still need to supply `-auto-approve=true` when running Terraform
commands that may make changes. We have seen that if Terraform prompts
for input in a Concourse task it will pause forever awaiting input,
and when killed will write an empty statefile. This would have to be
fixed manually using the previous version of the S3 file.

[1] https://www.terraform.io/docs/configuration/environment-variables.html#tf_input

We might want to override this in the future if we choose to use Terraform
outside of the pipeline (e.g. Pingdom setup) with the same container.