bosh-shell: handle non-existing CF deployment
Currently the startup script attempts to download the bosh manifest for
the corresponding CF deployment and target it. If this deployment
doesn't exist, it causes the download to fail, and the startup script to
abort. As a result it's not possible to use this when the CF deployment
doesn't exist.

This updates the script to gracefully handle a missing CF deployment so
that you can still interact with bosh.