Prepare the Bosh CLI v2 image to create an environment
We are migrating from CLI version v1 to v2 which means we want to use e.g.
`bosh create-env` instead of the `bosh-init` command.
The necessary dependencies are defined in [1] and to keep the maintenance
easier we change the base image to Ubuntu 16.04.

We updated the Bosh CLI to 2.0.48 so we can use the latest version.

A dummy `bosh create-env` command will run at the end of the build to cache
some resources to speed up the create process. This was copied from the
bosh-init Docker image.

[1] https://bosh.io/docs/cli-env-deps.html