Add image for executing arbitrary Bosh commands
This allows you to execute one-off commands using Bosh. The initial
motivation for this comes from the repository paas-infrastructure-map,
which provides the tools necessary to query Bosh and build a dynamic
map of a Bosh deployment.

We use the empty exec-style `CMD` declaration to override the command
of the base Ruby image.