Add log-cache plugin to the cf-acceptance-tests
We want to test the log-cache service using the cf-acceptance-tests[1],
but for that we need the corresponding plugin installed in the
container.

Additionally, as the tests would change the CF_HOME variable, we must
set the CF_PLUGIN_HOME at the container level so the containers
are available when the tests run.

The plugins can now be installed directly from a url in the internet,
so we simplify how the plugins are installed.

We also add tests for the plugins in the container.

[1] https://github.com/cloudfoundry/cf-acceptance-tests/commit/923a0fe9335bf0a1e75286801e952f248491cd38