Update bosh-cli for bosh and bosh-shell
We have update the bosh-cli-v2 to use bosh-cli 5.2.1 with SKI for
certs, and we want to use the same in the rest of containers
for consistency.

The version 5.2.1 [1] supports generating certificates via
--vars-store with SKI/AKI.

We need this to be able to rotate CA certificates used by CF API,
otherwise it is not possible to use two CA certs with the same name
at the same time.

More info in [2]

[1] https://github.com/cloudfoundry/bosh-cli/releases/tag/v5.2.1
[2] https://github.com/cloudfoundry/config-server/pull/9