Don't pass command line parameters to `bosh login`
Bosh is becoming more dependent on using UAA to authenticate, so we are
going to start using UAA. This means on every operation you have to pass
these flags, otherwise it can use the wrong OAuth flow. For example, the
initial `bosh login` may work, but a subsequent `bosh vms` operation
will result in authorization failure.

The correct usage of the Bosh CLI now is to use the environment
variables `BOSH_CLIENT` and `BOSH_CLIENT_SECRET`, which ensures the
correct authentication flow on every operation. We can pass said
variables when invoking this image, so don't need to reference them in
this script.