No longer do `bosh login` in bosh-shell entrypoint
Now we are passing the environment variables `BOSH_CLIENT` and
`BOSH_CLIENT_SECRET` we do not need to perform a login. The Bosh CLI
will negotiate authentication on every operation anyway. Removing this
has the added benefit of not failing the startup script if login fails,
which means you can inspect the contents of the container more easily.