Bump to alpine 3.9 and ruby to 2.6
Closes #153

This commit makes use of `apk add --no-cache` instead of `--update` then
rm.

This commit updates a bunch of tests to include the correct version

There is a toil-ey depedency chain between awscli and cf-cli which is
annoying so I removed the test that stops you updating them at the same
time.

Signed-off-by: Toby Lorne <toby.lornewelch-richards@digital.cabinet-office.gov.uk>