Add an image that contains middleman
We use https://middlemanapp.com/ to build things like the team manual
and the tech docs.

The team manual sometimes takes almost an hour to build:

https://concourse.build.ci.cloudpipeline.digital/teams/main/pipelines/paas-team-manual/jobs/publish/builds/95

Because it's doing an apt-get update / install every time it builds the docs.

If we had a better base image than ruby:2.5-slim (like one that already
had the other dependencies) we'd be able to skip the apt-get bits and
just jump in at `bundle install && bundle exec middleman build`.

For good measure I've installed middleman in this image directly, but in
practice we'd probably want the apps to be able to pick their own
version with bundler.