Allow non-interactive bosh-shell usage
Now that we configure credhub CLI as part of this container, we have reasonable
call to use it to access secrets in a non-interactive fashion as part of our
dev tooling. This PR feeds the command used to run the container through to the
bash shell started in `startup.sh`.

In order to correctly run a command, scripts should prefix their commands with
`-c`, as if they were running `bash -c 'command'" (which they are). E.g.

```
docker run \
  --rm \
  governmentpaas/bosh-shell:latest \
  "-c 'credhub --version'"
```

Interactive usage should continue to function as normal when the container is
run interactively without a command. E.g.

```
docker run -it --rm governmentpaas/bosh-shell:latest
```