ci(github): use github actions
We publish to docker hub from github as DockerHub autobuild is not free anymore.
We only build and test our customized images, not all images defined in upstream repository

We have disabled ghcr push. As org name are different between dockerhub and github, we cannot use the same namespace. So we do not need ghcr publication for now.

remove anchor, not supported yet

Error:
`The workflow is not valid. .github/workflows/docker-image.yml: Anchors are not currently supported. Remove the anchor 'publish_image_with'`