Allow passing secrets to the reusable workflow
Github actions is a bit arcane in some ways.

By default, secrets aren't passed directly to reusable workflows,
because otherwise you could end up passing secrets to a bad actor
if using a workflow from a third party.

Normally, we would have to specify the specific secrets that need to be
passed to the reusable workflow, but as it's a local one, we can just
specify that it inherits the provided secrets.

Note: I'm not sure why the GITHUB_TOKEN automatic secret doesn't cause
these same issues!