Workflow improvements
* Elevate GITHUB_TOKEN permissions
* Always build and push on PR (for testing built images)
* Automatically detect if there are acceptance tests
* Only lint on pull requests, not each commit

GITHUB_TOKEN changes are documented with comments, but the TL;DR is:

`actions: write`: This allows 'cancel-workflow-action' to cancel
existing running actions. If another action has started on a branch / pr
it's highly unlikely we care about the outputs from older ones.

`statuses: write`: This allows 'super-linter' to write to commit / pr
statuses, which gives an at-a-glance idea of tue results, without having
to scour the job output.